Another test added
There is a bug when you spam the client command with a client who is not the current client that causes the program to get stuck. The solution for everything to be perfect would be change the logic of the entire program so the client awaits a confirmation for every bit sended, getting rid of the usleep function. Not gonna change the entire program now :3